Stabilize serial ports tests on GHA (#190)
Serial port tests have been failing intermittently on GHA for what
looked like a delay in write operations. According to spec data should
be available right away, but for some reason it is not.

Update test to do more checks and reporting. This allowed to reproduce
issue locally in devcontainer
Add 100ms sleep between write and read operations, this seems to resolve
the issue for now both locally and GHA