A bunch of random fixes (#212)
Make venv isolated from system. It's purpose is to make docs, but with
system mpl visible build breaks.
Use system python interpreter as default. This results in colcon/cmake
to work fine and discover system ros python packages
Enable type checker for python and fix a bunch of type errors in the
production code
Add all the needed info to `config/drqp.yml` but it has not been tested