pyDualsense playeroo (#219)
Added a control script for ps5 controller that uses pydualsense which
uses hidapi under the hood and provides direct access to touch button
and touchpad touches inputs. The touchpad info always shows touch as
active though, so it is not really usable.
pydualsense depends on the libhidapi-dev
Added ansible role to install libhidapi-dev and configured udev rules to
allow access to hidraw to everyone